requirements.txt
Flask==0.10.1
Flask-PyMongo==0.5.1
Flask-Restless==0.13.1
Flask-SQLAlchemy==0.16
Jinja2==2.7
MarkupSafe==0.18
SQLAlchemy==0.8.2
Werkzeug==0.9.2
argparse==1.2.1
beautifulsoup4==4.5.3
click==6.7
gunicorn==17.5
itsdangerous==0.22
mimerender==0.5.4
pymongo==3.5.1
python-dateutil==2.1
python-dotenv==0.7.1
python-mimeparse==0.1.4
requests==1.2.3
six==1.3.0
wsgiref==0.1.2
